Removed Vundle as a sub-module.
Using Vundle to manage Vundle. Uh, it's a bit hard to explain. Basically
you run ./sync-bundles.sh and it will go and grab Vundle if it needs to.
It will then use Vundle to keep Vundle and all other bundles up to date.
Sub-modules have always seemed awkward and clunky to me.